+++ src/natcap/invest/routing/routedem.py
@@ -1,19 +1,13 @@
-"""RouteDEM entry point for exposing the natcap.invest's routing package
-    to a UI."""
-
+"""RouteDEM for exposing the natcap.invest's routing package to UI."""
 import os
 import logging
-
 
 import pygeoprocessing.geoprocessing
 import pygeoprocessing.routing
 import pygeoprocessing.routing.routing_core
 
-
-logging.basicConfig(format='%(asctime)s %(name)-20s %(levelname)-8s \
-%(message)s', level=logging.DEBUG, datefmt='%m/%d/%Y %H:%M:%S ')
-
 LOGGER = logging.getLogger('natcap.invest.routing.routedem')
+
 
 def execute(args):
     """RouteDEM: D-Infinity Routing.
@@ -61,48 +55,41 @@ def execute(args):
     Returns:
         ``None``
     """
-
     output_directory = args['workspace_dir']
     LOGGER.info('creating directory %s', output_directory)
-    if not os.path.exists(output_directory):
-        os.makedirs(output_directory)
-    file_suffix = ''
+    pygeoprocessing.create_directories([output_directory])
     dem_uri = args['dem_uri']
 
     LOGGER.info('resolving filling pits')
 
-    prefix, suffix = os.path.splitext(args['pit_filled_filename'])
     dem_tiled_uri = os.path.join(
-        output_directory, 'dem_tiled' + file_suffix + '.tif')
-    pygeoprocessing.geoprocessing.tile_dataset_uri(dem_uri, dem_tiled_uri, 256)
+        output_directory, 'dem_tiled.tif')
+    pygeoprocessing.geoprocessing.tile_dataset_uri(
+        dem_uri, dem_tiled_uri, 256)
     dem_pit_filled_uri = os.path.join(
-        output_directory, prefix + file_suffix + suffix)
+        output_directory, args['pit_filled_filename'])
     pygeoprocessing.routing.fill_pits(dem_tiled_uri, dem_pit_filled_uri)
     dem_uri = dem_pit_filled_uri
 
-    #Calculate slope
+    # Calculate slope
     if args['calculate_slope']:
         LOGGER.info("Calculating slope")
-        prefix, suffix = os.path.splitext(args['slope_filename'])
-        slope_uri = os.path.join(
-            output_directory, prefix + file_suffix + suffix)
+        slope_uri = os.path.join(output_directory, args['slope_filename'])
         pygeoprocessing.geoprocessing.calculate_slope(dem_uri, slope_uri)
 
-    #Calculate flow accumulation
+    # Calculate flow accumulation
     LOGGER.info("calculating flow direction")
-    prefix, suffix = os.path.splitext(args['flow_direction_filename'])
     flow_direction_uri = os.path.join(
-        output_directory, prefix + file_suffix + suffix)
+        output_directory, args['flow_direction_filename'])
     pygeoprocessing.routing.flow_direction_d_inf(dem_uri, flow_direction_uri)
 
     LOGGER.info("calculating flow accumulation")
-    prefix, suffix = os.path.splitext(args['flow_accumulation_filename'])
     flow_accumulation_uri = os.path.join(
-        output_directory, prefix + file_suffix + suffix)
+        output_directory, args['flow_accumulation_filename'])
     pygeoprocessing.routing.flow_accumulation(
         flow_direction_uri, dem_uri, flow_accumulation_uri)
 
-    #classify streams from the flow accumulation raster
+    # classify streams from the flow accumulation raster
     LOGGER.info("Classifying streams from flow accumulation raster")
 
     if args['multiple_stream_thresholds']:
@@ -116,21 +103,19 @@ def execute(args):
                 "Calculating stream threshold at %s pixels",
                 threshold_amount)
             v_stream_uri = os.path.join(
-                output_directory, 'v_stream%s_%s.tif' %
-                (file_suffix, str(threshold_amount)))
+                output_directory, 'v_stream_%s.tif' %
+                (str(threshold_amount),))
 
             pygeoprocessing.routing.stream_threshold(
                 flow_accumulation_uri, threshold_amount, v_stream_uri)
     else:
-        v_stream_uri = os.path.join(
-            output_directory, 'v_stream%s.tif' % file_suffix)
+        v_stream_uri = os.path.join(output_directory, 'v_stream.tif')
         pygeoprocessing.routing.stream_threshold(
             flow_accumulation_uri, float(args['threshold_flow_accumulation']),
             v_stream_uri)
 
     if args['calculate_downstream_distance']:
-        prefix, suffix = os.path.splitext(args['downstream_distance_filename'])
         distance_uri = os.path.join(
-            output_directory, prefix + file_suffix + suffix)
+            output_directory, args['downstream_distance_filename'])
         pygeoprocessing.routing.distance_to_stream(
             flow_direction_uri, v_stream_uri, distance_uri)